refactor: create macthCreator as separate library
