+++ mainwindow.ui
@@ -340,6 +340,7 @@
     <addaction name="actionSave_Player_List"/>
     <addaction name="actionSave_Schedule"/>
     <addaction name="actionSave_Points_Table"/>
+    <addaction name="actionSave_All_in_One"/>
    </widget>
    <widget class="QMenu" name="menuLanguage">
     <property name="title">
@@ -433,6 +434,15 @@
     <font>
      <pointsize>12</pointsize>
     </font>
+   </property>
+  </action>
+  <action name="actionSave_All_in_One">
+   <property name="icon">
+    <iconset resource="Images.qrc">
+     <normaloff>:/assets/save.png</normaloff>:/assets/save.png</iconset>
+   </property>
+   <property name="text">
+    <string>Save All in One</string>
    </property>
   </action>
  </widget>
