+++ src/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(backend)
+add_subdirectory(frontend)
